refactor: extract ORION engine into proximadb-orion-engine crate (ORION move 6g)
Capstone of the ORION extraction cascade: the 14 ORION engine files move from
src/graph/engines/orion/ into a new modality crate `proximadb-orion-engine`. A
root shim (`src/graph/engines/orion.rs`: `pub use proximadb_orion_engine::*`)
preserves every existing `crate::graph::engines::orion::*` path, so all callers
resolve unchanged. This completes PRs #1083/#1085/#1090/#1092/#1097/#1098-#1102
(the WAL + GraphEngine-trait + distance/filesystem/cache/canonical-WAL
decoupling) — orion/ now lives in its own crate with NO root-crate dependency.

Moved:
- 14 files (lib.rs, algorithms/, compaction, disk_storage, index, multi_tenant,
  persistence, storage, traversal) via `git mv` (history preserved).
- GraphMemoryPool (struct + impl + Default) from src/graph/mod.rs into the crate
  (memory_pool.rs); root re-exports it via the shim for its 5 existing callers.
- property_value_to_string (GraphMemoryPool index helper) into memory_pool.rs.

Rewrites in the moved files (crate:: -> leaf paths):
- graph/graph::model -> proximadb_graph_model; engines::GraphEngine ->
  proximadb_graph_engine_traits; storage::tenant -> proximadb_storage_tenant;
  proto -> proximadb_proto; memtable GraphOperation -> graph_model;
  core::serialization CompressionAlgorithm -> proximadb_compression_types;
  metrics::td066_metrics -> proximadb_metrics (control-tier; modality->control
  is allowed by the workspace boundaries); self-refs -> crate::.
- disk_storage UnifiedWALWriter stub field -> dyn GraphWalPort.
- canonical_replay_scope_enabled -> inlined env-read.

Tests: the TD-066 WAL-behavior tests (need the root unified_wal_factory)
extracted to a root module `src/graph/engines/orion_recovery_tests.rs`;
OrionGraphEngine.memory_pool made pub for test assertions. The engine's own
unit tests stay in the crate.

Crate wired into the workspace (members + workspace deps + root dep). Deps are
foundation/storage/control/modality leaves only — no upward layering.

Verified: cargo clippy --lib --bins (CI gate) + cargo check --tests clean;
rustfmt 1.95.0; panic-policy, layering, workspace-boundary guards clean (no
boundary findings).
