+++ crates/modalities/proximadb-orion-engine/Cargo.toml
@@ -1,0 +1,51 @@
+[package]
+name = "proximadb-orion-engine"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ORION graph engine: in-memory CSR traversal + WAL-persisted topology projection over canonical records. Extracted from the root crate (ORION move 6g)."
+
+[lib]
+name = "proximadb_orion_engine"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+bincode.workspace = true
+chrono.workspace = true
+dashmap.workspace = true
+lz4_flex.workspace = true
+lru = "0.18"
+memmap2 = "0.9"
+parking_lot.workspace = true
+rand.workspace = true
+rayon.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+snap.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+zstd.workspace = true
+# Foundation-tier leaf deps (graph model/contract, kernel, distance, compression, proto)
+proximadb-compression-types.workspace = true
+proximadb-distance-kernel.workspace = true
+proximadb-graph-engine-traits.workspace = true
+proximadb-graph-model.workspace = true
+proximadb-kernel.workspace = true
+proximadb-proto.workspace = true
+# Storage-tier leaf deps (WAL/filesystem/cache ports + tenant + common WAL types)
+proximadb-storage-common.workspace = true
+proximadb-storage-ports.workspace = true
+proximadb-storage-tenant.workspace = true
+# Control-tier (modality -> control is allowed by the workspace boundaries)
+proximadb-metrics.workspace = true
+# Modality-tier: graph algorithms/compaction (re-exported by the engine's shim modules)
+proximadb-graph = { path = "../proximadb-graph" }
+
+[lints]
+workspace = true
