+++ crates/modalities/proximadb-orion-engine/src/algorithms/traits.rs
